chore(EMS-2206): fix/update eligibility summary list answer
