+++ src/ui/server/helpers/summary-lists/eligibility/eligibility-fields/index.test.ts
@@ -1,5 +1,6 @@
 import generateEligibilityFields from '.';
 import { FIELDS_ELIGIBILITY } from '../../../../content-strings/fields/insurance';
+import { TOTAL_CONTRACT_VALUE } from '../../../../constants';
 import INSURANCE_FIELD_IDS from '../../../../constants/field-ids/insurance';
 import fieldGroupItem from '../../generate-field-group-item';
 import getFieldById from '../../../get-field-by-id';
@@ -10,6 +11,7 @@ const { ELIGIBILITY: FIELD_IDS } = INSURANCE_FIELD_IDS;
 
 const {
   WANT_COVER_OVER_MAX_AMOUNT,
+  TOTAL_CONTRACT_VALUE: TOTAL_CONTRACT_VALUE_ELIGIBILITY,
   WANT_COVER_OVER_MAX_PERIOD,
   OTHER_PARTIES_INVOLVED,
   LETTER_OF_CREDIT,
@@ -19,6 +21,8 @@ const {
   HAS_MINIMUM_UK_GOODS_OR_SERVICES,
   VALID_EXPORTER_LOCATION,
 } = FIELD_IDS;
+
+const { MORE_THAN_500K } = TOTAL_CONTRACT_VALUE;
 
 describe('server/helpers/summary-lists/eligibility/eligibility-fields', () => {
   const mockAnswers = mockApplication.eligibility;
@@ -54,7 +58,7 @@ describe('server/helpers/summary-lists/eligibility/eligibility-fields', () => {
         data: mockAnswers,
         renderChangeLink: false,
       },
-      mapYesNoField(mockAnswers[WANT_COVER_OVER_MAX_AMOUNT]),
+      mapYesNoField(mockAnswers[TOTAL_CONTRACT_VALUE_ELIGIBILITY].valueId === MORE_THAN_500K.DB_ID),
     ),
     fieldGroupItem(
       {
