+++ src/ui/server/helpers/summary-lists/eligibility/eligibility-fields/index.ts
@@ -1,4 +1,5 @@
 import { FIELDS_ELIGIBILITY } from '../../../../content-strings/fields/insurance';
+import { TOTAL_CONTRACT_VALUE } from '../../../../constants';
 import INSURANCE_FIELD_IDS from '../../../../constants/field-ids/insurance';
 import fieldGroupItem from '../../generate-field-group-item';
 import getFieldById from '../../../get-field-by-id';
@@ -9,6 +10,7 @@ const { ELIGIBILITY: FIELD_IDS } = INSURANCE_FIELD_IDS;
 
 const {
   WANT_COVER_OVER_MAX_AMOUNT,
+  TOTAL_CONTRACT_VALUE: TOTAL_CONTRACT_VALUE_ELIGIBILITY,
   WANT_COVER_OVER_MAX_PERIOD,
   OTHER_PARTIES_INVOLVED,
   LETTER_OF_CREDIT,
@@ -18,6 +20,8 @@ const {
   HAS_MINIMUM_UK_GOODS_OR_SERVICES,
   VALID_EXPORTER_LOCATION,
 } = FIELD_IDS;
+
+const { MORE_THAN_500K } = TOTAL_CONTRACT_VALUE;
 
 /**
  * generateEligibilityFields
@@ -57,7 +61,7 @@ const generateEligibilityFields = (answers: InsuranceEligibility) => {
         data: answers,
         renderChangeLink: false,
       },
-      mapYesNoField(answers[WANT_COVER_OVER_MAX_AMOUNT]),
+      mapYesNoField(answers[TOTAL_CONTRACT_VALUE_ELIGIBILITY].valueId === MORE_THAN_500K.DB_ID),
     ),
     fieldGroupItem(
       {
