+++ src/ui/server/test-mocks/mock-eligibility.ts
@@ -1,4 +1,5 @@
 import mockCountries from './mock-countries';
+import { TOTAL_CONTRACT_VALUE } from '../constants';
 import { InsuranceEligibility } from '../../types';
 
 const mockEligibility = {
@@ -13,7 +14,7 @@ const mockEligibility = {
   wantCoverOverMaxPeriod: false,
   totalContractValue: {
     value: 'Mock value',
-    valueId: 1,
+    valueId: TOTAL_CONTRACT_VALUE.LESS_THAN_500K.DB_ID,
   },
 } as InsuranceEligibility;
